agents/requirements.txt
fastapi
uvicorn[standard]
google-adk
vertexai
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
python-dotenv
google-auth
PyMuPDF
requests
beautifulsoup4
tavily-python
aiohttp